requirements.txt
spacy>=3.5.0
PyPDF2>=3.0.0
python-docx>=0.8.11
pdfminer.six>=20221105
beautifulsoup4>=4.11.1
pandas>=1.5.0
scikit-learn>=1.2.0
tqdm>=4.64.1
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
aiofiles>=0.8.0
regex>=2022.10.31
pytest>=7.3.1
pytest-asyncio>=0.21.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl
